gdfg

+++ Assets/Scripts/Item/Items/propeller.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class propeller : Item
 {
-    [SerializeField] int burstTime;
+    [SerializeField] float burstTime;
     public override void Use()
     {
         playerUser.BurstAcelerate(burstTime);
